pyproject.toml
[project]
name = "mct"
version = "0.1.2"
authors = [{ name = "Oscar Colunga", email = "oscar@ancile.dev" }]
description = "macOS Configuration Tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["typer>=0.9.0", "pyyaml>=6.0"]

[dependency-groups]
dev = [
 "ruff>=0.11.4",
 "ty>=0.0.12",
]

[project.scripts]
mct = "mct.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
